updated version, included repair function, updated presets and updatertarget

+++ GUI/src/TalentTreeEditorWindow.cpp
@@ -40,7 +40,10 @@ namespace TTM {
                 ImGui::Text(talent->getName().c_str());
                 ImGui::PopFont();
                 if (uiData.iconIndexMap.count(talent->index)) {
-                    ImGui::Image(uiData.iconIndexMap.at(talent->index).first.texture, ImVec2(uiData.treeEditorBaseTalentSize, uiData.treeEditorBaseTalentSize));
+                    ImGui::Image(
+                        uiData.iconIndexMap.at(talent->index).first.texture, 
+                        ImVec2(static_cast<float>(uiData.treeEditorBaseTalentSize), static_cast<float>(uiData.treeEditorBaseTalentSize))
+                    );
                 }
                 //ImGui::SameLine(ImGui::GetContentRegionAvail().x - ImGui::CalcTextSize(idLabel.c_str()).x);
                 ImGui::Text(idLabel.c_str());
@@ -73,7 +76,10 @@ namespace TTM {
                 ImGui::Text(talent->name.c_str());
                 ImGui::PopFont();
                 if (uiData.iconIndexMap.count(talent->index)) {
-                    ImGui::Image(uiData.iconIndexMap.at(talent->index).first.texture, ImVec2(uiData.treeEditorBaseTalentSize, uiData.treeEditorBaseTalentSize));
+                    ImGui::Image(
+                        uiData.iconIndexMap.at(talent->index).first.texture, 
+                        ImVec2(static_cast<float>(uiData.treeEditorBaseTalentSize), static_cast<float>(uiData.treeEditorBaseTalentSize))
+                    );
                 }
                 //ImGui::SameLine(ImGui::GetContentRegionAvail().x - ImGui::CalcTextSize(idLabel.c_str()).x);
                 ImGui::Text(idLabel.c_str());
@@ -94,7 +100,10 @@ namespace TTM {
                 ImGui::Text(talent->nameSwitch.c_str());
                 ImGui::PopFont();
                 if (uiData.iconIndexMap.count(talent->index)) {
-                    ImGui::Image(uiData.iconIndexMap.at(talent->index).second.texture, ImVec2(uiData.treeEditorBaseTalentSize, uiData.treeEditorBaseTalentSize));
+                    ImGui::Image(
+                        uiData.iconIndexMap.at(talent->index).second.texture,
+                        ImVec2(static_cast<float>(uiData.treeEditorBaseTalentSize), static_cast<float>(uiData.treeEditorBaseTalentSize))
+                    );
                 }
                 //ImGui::SameLine(ImGui::GetContentRegionAvail().x - ImGui::CalcTextSize(idLabel.c_str()).x);
                 ImGui::Text(idLabel.c_str());
